+++ examples/chapi-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @learncard/chapi-example
 
+## 1.0.88
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.21
+    -   @learncard/react@2.6.26
+
+## 1.0.87
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.20
+    -   @learncard/react@2.6.25
+
 ## 1.0.86
 
 ### Patch Changes