+++ packages/plugins/learn-cloud/CHANGELOG.md
@@ -1,5 +1,12 @@
 # learn-card-core
 
+## 1.1.12
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/learn-cloud-client@1.1.12
+
 ## 1.1.11
 
 ### Patch Changes