+++ services/learn-card-discord-bot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # learn-card-discord-bot
 
+## 1.1.48
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.21
+
+## 1.1.47
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.20
+
 ## 1.1.46
 
 ### Patch Changes